+++ NEWS.md
@@ -1,3 +1,13 @@
+# 2020-02-06
+
+* bug in statistic_override with functions
+* bug caused by upstream changes in tab_style
+* bug caused by upstream changes in filename='rtf'
+* Allow multiple rows of uncertainty estimates per coefficient
+* Preserve add_rows order
+* Display uncertainty estimates next to the coefficient with statistic_vertical = FALSE
+* Better clean_latex function
+
 # 2019-06-30
 
-Initial release (gt still needs to be installed from github)
+* Initial release (gt still needs to be installed from github)